refactor(ci-metrics): move CI logic to library module to fix Next.js route build validation
